fix: icon rename
fix: icon rename
