+++ services/ark-dashboard/ark-dashboard/components/icons/arrow-back.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+
+import { cn } from '@/lib/utils';
+
+interface IconProps extends React.SVGProps<SVGSVGElement> {
+  readonly className?: string;
+}
+
+export function ArrowBack({ className, ...props }: Readonly<IconProps>) {
+  return (
+    <svg
+      className={cn('', className)}
+      viewBox="0 -960 960 960"
+      fill="currentColor"
+      aria-hidden="true"
+      {...props}>
+      <path d="M266.31-457.31 512-212l-32 32-300-300 300-300 32 32-245.69 245.31H780v45.38H266.31Z" />
+    </svg>
+  );
+}
