+++ services/ark-dashboard/ark-dashboard/components/icons/arrow-upward-alt.tsx
