+++ services/ark-dashboard/ark-dashboard/components/icons/bar-chart.tsx
