+++ services/ark-dashboard/ark-dashboard/components/icons/cancel.tsx
