+++ services/ark-dashboard/ark-dashboard/components/icons/earthquake.tsx
