+++ services/ark-dashboard/ark-dashboard/components/icons/error-icon.tsx
