+++ services/ark-dashboard/ark-dashboard/components/icons/index.ts
@@ -1,48 +1,49 @@
-export { AccountTree } from './AccountTree';
-export { Add } from './Add';
-export { ArrowDownwardAlt } from './ArrowDownwardAlt';
-export { ArrowForward } from './ArrowForward';
-export { ArrowUpwardAlt } from './ArrowUpwardAlt';
-export { BarChart } from './BarChart';
-export { Bedtime } from './Bedtime';
-export { CalendarMonth } from './CalendarMonth';
-export { Cancel } from './Cancel';
-export { ChatBubble } from './ChatBubble';
-export { Check } from './Check';
-export { CheckCircle } from './CheckCircle';
-export { ChevronDown } from './ChevronDown';
-export { ChevronLeft } from './ChevronLeft';
-export { ChevronRight } from './ChevronRight';
-export { Close } from './Close';
-export { Dashboard } from './Dashboard';
-export { Database } from './Database';
-export { Dns } from './Dns';
-export { Earthquake } from './Earthquake';
-export { Edit } from './Edit';
-export { ErrorIcon } from './ErrorIcon';
-export { FilterAlt } from './FilterAlt';
-export { Group } from './Group';
-export { Handyman } from './Handyman';
-export { Help } from './Help';
-export { Info } from './Info';
-export { InsertDriveFile } from './InsertDriveFile';
-export { KeyboardDoubleArrowLeft } from './KeyboardDoubleArrowLeft';
-export { KeyboardDoubleArrowRight } from './KeyboardDoubleArrowRight';
-export { LightMode } from './LightMode';
-export { Memory } from './Memory';
-export { PlaylistAddCheck } from './PlaylistAddCheck';
-export { PlugConnect } from './PlugConnect';
-export { Plus } from './Plus';
-export { SaveAlt } from './SaveAlt';
-export { Schedule } from './Schedule';
-export { Search } from './Search';
-export { Send } from './Send';
-export { Settings } from './Settings';
-export { Shield } from './Shield';
-export { SingleTool } from './SingleTool';
-export { SmartToy } from './SmartToy';
-export { Storefront } from './Storefront';
-export { SwapVert } from './SwapVert';
-export { Trash } from './Trash';
-export { UnfoldMore } from './UnfoldMore';
-export { VpnKey } from './VpnKey';
+export { AccountTree } from './account-tree';
+export { Add } from './add';
+export { ArrowBack } from './arrow-back';
+export { ArrowDownwardAlt } from './arrow-downward-alt';
+export { ArrowForward } from './arrow-forward';
+export { ArrowUpwardAlt } from './arrow-upward-alt';
+export { BarChart } from './bar-chart';
+export { Bedtime } from './bedtime';
+export { CalendarMonth } from './calendar-month';
+export { Cancel } from './cancel';
+export { ChatBubble } from './chat-bubble';
+export { Check } from './check';
+export { CheckCircle } from './check-circle';
+export { ChevronDown } from './chevron-down';
+export { ChevronLeft } from './chevron-left';
+export { ChevronRight } from './chevron-right';
+export { Close } from './close';
+export { Dashboard } from './dashboard';
+export { Database } from './database';
+export { Dns } from './dns';
+export { Earthquake } from './earthquake';
+export { Edit } from './edit';
+export { ErrorIcon } from './error-icon';
+export { FilterAlt } from './filter-alt';
+export { Group } from './group';
+export { Handyman } from './handyman';
+export { Help } from './help';
+export { Info } from './info';
+export { InsertDriveFile } from './insert-drive-file';
+export { KeyboardDoubleArrowLeft } from './keyboard-double-arrow-left';
+export { KeyboardDoubleArrowRight } from './keyboard-double-arrow-right';
+export { LightMode } from './light-mode';
+export { Memory } from './memory';
+export { PlaylistAddCheck } from './playlist-add-check';
+export { PlugConnect } from './plug-connect';
+export { Plus } from './plus';
+export { SaveAlt } from './save-alt';
+export { Schedule } from './schedule';
+export { Search } from './search';
+export { Send } from './send';
+export { Settings } from './settings';
+export { Shield } from './shield';
+export { SingleTool } from './single-tool';
+export { SmartToy } from './smart-toy';
+export { Storefront } from './storefront';
+export { SwapVert } from './swap-vert';
+export { Trash } from './trash';
+export { UnfoldMore } from './unfold-more';
+export { VpnKey } from './vpn-key';
