+++ services/ark-dashboard/ark-dashboard/components/icons/keyboard-double-arrow-right.tsx
