+++ services/ark-dashboard/ark-dashboard/components/icons/light-mode.tsx
