+++ services/ark-dashboard/ark-dashboard/components/icons/memory.tsx
