+++ services/ark-dashboard/ark-dashboard/components/icons/plug-connect.tsx
