+++ services/ark-dashboard/ark-dashboard/components/icons/settings.tsx
