+++ services/ark-dashboard/ark-dashboard/components/icons/shield.tsx
