+++ services/ark-dashboard/ark-dashboard/components/icons/smart-toy.tsx
