+++ services/ark-dashboard/ark-dashboard/components/icons/trash.tsx
