+++ services/ark-dashboard/ark-dashboard/components/sessions-conversations/new-session-dialog.tsx
@@ -14,8 +14,8 @@ import { Checkbox } from '@/components/ui/checkbox';
 import { TagToggle } from '@/components/ui/tag-toggle';
 import { Separator } from '@/components/ui/separator';
 import { IconShell } from '@/components/ui/icon-shell';
-import { Search } from '@/components/icons/Search';
-import { Close } from '@/components/icons/Close';
+import { Search } from '@/components/icons/search';
+import { Close } from '@/components/icons/close';
 import { useQuery } from '@tanstack/react-query';
 import { useNamespacedNavigation } from '@/lib/hooks/use-namespaced-navigation';
 import { agentsService } from '@/lib/services/agents';
