+++ services/ark-dashboard/ark-dashboard/components/ui/calendar.tsx
@@ -28,8 +28,8 @@ import {
 } from 'react-day-picker';
 
 import { cn } from '../../lib/utils';
-import { ChevronLeft } from '../icons/ChevronLeft';
-import { ChevronRight } from '../icons/ChevronRight';
+import { ChevronLeft } from '../icons/chevron-left';
+import { ChevronRight } from '../icons/chevron-right';
 import { Button } from './button';
 import { IconShell } from './icon-shell';
 import { Input } from './input';
