+++ services/ark-dashboard/ark-dashboard/components/ui/tag.tsx
@@ -2,7 +2,7 @@ import { type VariantProps, cva } from 'class-variance-authority';
 import * as React from 'react';
 
 import { cn } from '../../lib/utils';
-import { Close } from '../icons/Close';
+import { Close } from '../icons/close';
 
 const hoverOverlay = {
   normal:
